Move Cloud Run OpenTelemetry plugin to temporal-gcp-cloud-run-opentelemetry
Rename the contrib/temporal-gcp-cloud-run module to contrib/temporal-gcp-cloud-run-opentelemetry and nest its source under the io.temporal.gcp.cloudrun.opentelemetry package, so the OpenTelemetry plugin sits alongside the sibling temporal-gcp-cloud-run-worker-id module. Update settings.gradle, the BOM, and the README to the new coordinates.
