+++ contrib/temporal-gcp-cloud-run-opentelemetry/src/main/java/io/temporal/gcp/cloudrun/opentelemetry/CloudRunOpenTelemetryPlugin.java
@@ -1,4 +1,4 @@
-package io.temporal.gcp.cloudrun;
+package io.temporal.gcp.cloudrun.opentelemetry;
 
 import io.opentelemetry.api.OpenTelemetry;
 import io.temporal.client.WorkflowClientOptions;
