+++ contrib/temporal-gcp-cloud-run-opentelemetry/src/test/java/io/temporal/gcp/cloudrun/opentelemetry/CloudRunOpenTelemetryPluginTest.java
@@ -1,4 +1,4 @@
-package io.temporal.gcp.cloudrun;
+package io.temporal.gcp.cloudrun.opentelemetry;
 
 import static org.junit.Assert.*;
