+++ temporal-bom/build.gradle
@@ -10,7 +10,7 @@ dependencies {
         api project(':temporal-opentelemetry')
         api project(':temporal-opentracing')
         api project(':temporal-aws-lambda')
-        api project(':temporal-gcp-cloud-run')
+        api project(':temporal-gcp-cloud-run-opentelemetry')
         api project(':temporal-remote-data-encoder')
         api project(':temporal-sdk')
         api project(':temporal-serviceclient')
